+++ src/app/page.tsx
@@ -1,8 +1,8 @@
 // /pages.tsx
-import MainHeader from "@/components/layout/MainHeader";
-import FooterNav from "@/components/layout/MainFooterNav";
-import MenuBar from "@/components/MenuBar";
-import Banner from "@/components/Banner";
+import MainHeader from "@/components/header/MainHeader";
+import FooterNav from "@/components/navbar/MainFooterNav";
+import MenuBar from "@/components/header/MenuBar";
+import Banner from "@/components/main/Banner";
 
 export default function Home() {
   return (
