+++ src/components/items/MainItem.tsx
